+++ main.go
@@ -12,6 +12,7 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
+	"path/filepath"
 	"syscall"
 	"time"
 
@@ -57,30 +58,31 @@ func main() {
 
 	config, err := getConfig()
 	if err != nil {
-		log.Print(err)
-		os.Exit(1)
+		log.Fatal(err)
 	}
 
 	log.Println("Cache time for /search: ", config.CacheTimeSearch)
 	log.Println("Cache time for /categories: ", config.CacheTimeCategories)
 	log.Println("Cache time for all others: ", config.CacheTimeCatchAll)
 
-	packagesBasePath := config.PublicDir + "/" + packageDir
-
-	// Prefill the package cache
+	packagesBasePath := filepath.Join(config.PublicDir, packageDir)
 	packages, err := util.GetPackages(packagesBasePath)
 	if err != nil {
-		log.Print(err)
-		os.Exit(1)
+		log.Fatal(err)
 	}
+
+	if len(packages) == 0 {
+		log.Fatal("No packages available")
+	}
+
 	log.Printf("%v package manifests loaded into memory.\n", len(packages))
 
 	server := &http.Server{Addr: address, Handler: getRouter(*config, packagesBasePath)}
 
 	go func() {
 		err := server.ListenAndServe()
-		if err != nil {
-			log.Printf("Error serving: %s", err)
+		if err != nil && err != http.ErrServerClosed {
+			log.Fatalf("Error occurred while serving: %s", err)
 		}
 	}()
 
@@ -90,7 +92,7 @@ func main() {
 
 	ctx := context.TODO()
 	if err := server.Shutdown(ctx); err != nil {
-		log.Print(err)
+		log.Fatal(err)
 	}
 }