+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Bugfixes
 
 ### Added
+* Use filepath.Walk to find valid package content data. [#438](https://github.com/elastic/package-registry/pull/438)
 
 ### Deprecated